meta setup.py since setuptools cant deal with the copy
